HUGE COMMIT
Currently, the webapp version does not work, but I believe that to be a problem with the remote phpmyadmin, coupled with a potential (theoretical) bug with the cdateofbirth being a date on the database, and a timestamp being handled in the app.

Major work (complete revamping) of the Search.java class. Made it work with the new and imporved search page.

Chuggin along...